MOSIP-25236 adhoc ui changes
