+++ src/components/FormAction.js
@@ -5,7 +5,7 @@ export default function FormAction({
   disabled = false,
 }) {
   const className =
-    "flex justify-center w-full font-medium rounded-lg text-sm px-5 py-2.5 text-center border border-2 ";
+    "flex justify-center w-full font-medium rounded-lg text-sm px-5 py-2 text-center border border-2 ";
 
   return (
     <>
